Fixed a minor bug in the program
Squashed the bug where when someone was inputting their Christmas code, they could potentially receive their own name, which was a cause of carelessness, as the programming was definitely there, it was just implemented in the wrong part, which caused it to fail. This change is further explained in the wiki.